cnn_image_classification/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
Pillow
numpy
scikit-learn
matplotlib
